Video issue fixed

+++ cypress.config.js
@@ -2,18 +2,16 @@ const { defineConfig } = require("cypress");
 const fs = require('fs');
 
 module.exports = defineConfig({
-  
   reporter: "mochawesome",
   reporterOptions: {
     reportDir: "cypress/reports",
     overwrite: false,
-    html: false,
-    json: true
+    html: true,  // Ensure HTML report is generated
+    json: true,
   },
 
   video: true, // Enable video recording
-  
-  
+
   env: {
     URL: "https://naveenautomationlabs.com/opencart/index.php?route=account/register",
   },
@@ -22,27 +20,28 @@ module.exports = defineConfig({
 
   e2e: {
     setupNodeEvents(on, config) {
-      // implement node event listeners here
+      // Delete video for passed tests
+      on('after:spec', (spec, results) => {  
+        if (results && results.video) {
+          // Check if all tests eventually passed    
+          const allEventuallyPassed = results.tests.every((test) =>
+            test.attempts.some((attempt) => attempt.state === 'passed')
+          );
+          
+          // If all tests passed, delete the video
+          if (allEventuallyPassed) {
+            fs.unlink(results.video, (err) => {
+              if (err) {
+                console.error('Failed to delete video:', err);
+              } else {
+                console.log('Deleted video:', results.video);
+              }
+            });
+          }
+        }
+      });
 
-       // Delete video for passed tests
-       on('after:spec', (spec, results) => {  if (results && results.video) 
-        {    // Check if all tests eventually passed    
-          const allEventuallyPassed = results.tests.every((test)=>test.attempts.some((attempt) => attempt.state === 'passed'))
-              if (allEventuallyPassed) 
-                { // Delete the video if all tests eventually passed
-                        fs.unlink(results.video, (err) => 
-                          {        
-                            if (err) 
-                              {         
-                                 console.error('Failed to delete video:', err)        
-                                } else 
-                                {          
-                                  console.log('Deleted video:', results.video)       
-                                 }      
-                                })    
-                              }  
-                            }
-                          })
+      // Required for Mochawesome reporter
       require('cypress-mochawesome-reporter/plugin')(on);
     },
   },
